Clean up warnings, imports and code between features
