+++ src/proof_system/widget/logic/verifierkey.rs
@@ -4,12 +4,17 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
-use super::{delta, delta_xor_and};
 use crate::commitment_scheme::kzg10::Commitment;
-use crate::proof_system::linearisation_poly::ProofEvaluations;
-#[cfg(feature = "alloc")]
-use alloc::vec::Vec;
-use dusk_bls12_381::{BlsScalar, G1Affine};
+
+cfg_if::cfg_if!(
+    if #[cfg(feature = "alloc")]
+    {
+        use crate::proof_system::linearisation_poly::ProofEvaluations;
+        use alloc::vec::Vec;
+        use dusk_bls12_381::{BlsScalar, G1Affine};
+        use super::proverkey::{delta, delta_xor_and};
+    }
+);
 
 #[derive(Debug, PartialEq, Eq, Copy, Clone)]
 pub(crate) struct VerifierKey {
