+++ src/proof_system/widget/permutation/verifierkey.rs
@@ -7,11 +7,16 @@
 #![allow(clippy::too_many_arguments)]
 
 use crate::commitment_scheme::kzg10::Commitment;
-use crate::permutation::constants::{K1, K2, K3};
-use crate::proof_system::linearisation_poly::ProofEvaluations;
-#[cfg(feature = "alloc")]
-use alloc::vec::Vec;
-use dusk_bls12_381::{BlsScalar, G1Affine};
+
+cfg_if::cfg_if!(
+    if #[cfg(feature = "alloc")]
+    {
+        use crate::proof_system::linearisation_poly::ProofEvaluations;
+        use alloc::vec::Vec;
+        use dusk_bls12_381::{BlsScalar, G1Affine};
+        use crate::permutation::constants::{K1, K2, K3};
+    }
+);
 
 #[derive(Debug, PartialEq, Eq, Copy, Clone)]
 pub(crate) struct VerifierKey {
