+++ src/proof_system/widget/range/mod.rs
@@ -8,15 +8,6 @@
 mod proverkey;
 mod verifierkey;
 
-use dusk_bls12_381::BlsScalar;
 #[cfg(feature = "alloc")]
 pub(crate) use proverkey::ProverKey;
 pub(crate) use verifierkey::VerifierKey;
-
-// Computes f(f-1)(f-2)(f-3)
-fn delta(f: BlsScalar) -> BlsScalar {
-    let f_1 = f - BlsScalar::one();
-    let f_2 = f - BlsScalar::from(2);
-    let f_3 = f - BlsScalar::from(3);
-    f * f_1 * f_2 * f_3
-}
